requirements.txt
django==4.2
python-decouple
psycopg[binary]
Pillow
drf-yasg
djangorestframework
pygammon